requirements.txt
discord.py-self==1.9.2
python-dotenv==1.0.0